Implement timeseries deprecations scheduled for v1.0
