+++ timeseries/src/autogluon/timeseries/__init__.py
@@ -6,10 +6,9 @@ except ImportError:
     pass
 
 from .dataset import TimeSeriesDataFrame
-from .evaluator import TimeSeriesEvaluator
 from .predictor import TimeSeriesPredictor
 
 _add_stream_handler()
 
 
-__all__ = ["TimeSeriesDataFrame", "TimeSeriesEvaluator", "TimeSeriesPredictor"]
+__all__ = ["TimeSeriesDataFrame", "TimeSeriesPredictor"]
