+++ timeseries/src/autogluon/timeseries/trainer/abstract_trainer.py
@@ -23,9 +23,9 @@ from autogluon.timeseries.metrics import TimeSeriesScorer, check_get_evaluation_
 from autogluon.timeseries.models.abstract import AbstractTimeSeriesModel
 from autogluon.timeseries.models.ensemble import AbstractTimeSeriesEnsembleModel, TimeSeriesGreedyEnsemble
 from autogluon.timeseries.models.presets import contains_searchspace
-from autogluon.timeseries.splitter import AbstractWindowSplitter, ExpandingWindowSplitter
 from autogluon.timeseries.utils.features import CovariateMetadata
 from autogluon.timeseries.utils.warning_filters import disable_tqdm
+from autogluon.timeseries.val_splitter import AbstractWindowSplitter, ExpandingWindowSplitter
 
 logger = logging.getLogger("autogluon.timeseries.trainer")
